make topology signature include a compressed list of available PUs
When mpirun collects topologies it first asks for signatures and then
only asks a host to send its topologif the signature is something new.
But with heterogeneous cgroups the signature doesn't have enough data
to notice differences that matter.

An example of the previous signature on one of our machines is
    2N:2S:22L3:22L2:44L1:44C:176H:ppc64le:le
which is just a count of how many numa nodes, sockets, cores, PUs etc
the host has.  If I create a small cgroup with just 4 PUs from one of
the cores, an example of the new signature is
    1N:1S:1L3:1L2:1L1:1C:4H:ppc64le:le

But if I did that on two different machines and used different cores
on each, the signature would look the same. One way to distinguish
such signatures is to also list the available PUs by os_index (I might
be okay with logical indexes if it was logical under a WHOLE-SYSTEM
topology, but in the non-WHOLE-SYSTEM topologies we're using  I don't
think the logical indexes would be good enough).

I think we should include a compression routine that recognizes
more patterns, but since this is a release branch I'm just using
    hwloc_bitmap_list_asprintf()
I've left in some commented out code I'd propose for future
consideration for string compression.  Patterns like filling the
second half of each socket might make hex masks like f0f0f0f0f0f0f0f0
which the above wouldn't compress by much, that would still be a
longish list of 5-8,13-16,21-24,...

signature strings end up things like
    2N:2S:22L3:22L2:44L1:44C:176H:H0-175:ppc64le:le
here "0-175" is the hwloc_bitmap_list_asprintf output.

Signed-off-by: Mark Allen <markalle@us.ibm.com>