opal/util: Include net/if.h to prevent macro mismatch
In cases where opal/util/if.h is included but net/if.h is not,
IF_NAMESIZE will be 32. If net/if.h is included on Linux systems,
IF_NAMESIZE will be 16. This can cause a mismatch when using the same
macro macro on a system. To avoid this error case, we include this file
in util/if.h.

Signed-off-by: William Zhang <wilzhang@amazon.com>