common_ompio_file_open: logic for communicator management
the logic for when to dup a communicator and when to leave it
was exactly the opposite of what we inteded to do: dup upon the 'real' file open,
do not dup on the file_open called from the sharedfp component.

This has no impact for most codes, but could have some impact on
multi-threaded executions.

Fixes issue #6963

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>