Rename VERSION to VERSION.txt
libc++ v8.0.0 and above uses a header named <version> in the C++
standard headers. VERSION file in this repo root conflicts with the
header from libc++ on case-insensitive file systems when building
the C++ bindings because the repo root is added to the search
directories by the build system.

Fixes https://github.com/open-mpi/ompi/issues/7155

Signed-off-by: Isuru Fernando <isuruf@gmail.com>