Check the return from "chdir" to avoid infinite loop
When autogen attempts to change to a new directory while processing a
subdirectory, it can get into an infinite loop if that directory
doesn't exist as it will remain in the top-level directory, see itself
there (as "autogen.pl"), and re-execute itself. Check the return code on
"chdir" and error out if it fails.

Signed-off-by: Ralph Castain <rhc@pmix.org>