ompio: fix abstraction violation in 3.1.x series
This commit is the minimum set required to fix the abstraction
violation in the ompio components for the 3.1.x series. This commit
does not change names of the structues (as it was done in the masters
and 4.x series), does also not address the compilation part (i.e.,
including headers of components). It should, however, take care of the
runtime loading problem of data from different components.

It is, to some extent, a simplified subset of pull request 5263.

Fixes issue https://github.com/open-mpi/ompi/issues/7309

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>
Signed-off-by: Jeff Squyres <jsquyres@cisco.com>