Remove the pml check_selected code.
The current logic is broken - this only works from the
node which has rank 0 on it. All other processes will fail
trying to fetch the data from rank 0 on a PMIx_Get().

Since this has been broken for what is probably a long time with no
complaints, it is probably safe to remove completely
and gain back some of the lost performance.

Fixes #7475

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>