configure: fix PRRTE conflict with -iquote
PRRTE needs hwloc and libevent, so it needs to be setup "late" in
configure.ac.  However, we don't want to do it at the absolute bottom
of configure.ac, because right near the bottom, we setup CPPFLAGS (and
others) with values that are expected to be used only in
Makefile[.am]'s -- i.e., "$(foo)" values.  Such values are not able to
be used here in configure.

Hence, move the PRRTE setup up above where we do these "final"/
only-relevant-to-Makefile[.am] CPPFLAGS (etc.) updates occur.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>